fix: logout to scan not otp
