+++ src/components/CampaignInitialisation/CampaignInitialisationScreen.tsx
@@ -273,6 +273,7 @@ export const CampaignInitialisationScreen: FunctionComponent<CampaignInitialisat
           component={MerchantPayoutStack}
         />
       )}
+      <Drawer.Screen name="LoadingScreen" component={LoadingView} />
     </Drawer.Navigator>
   ) : (
     <View style={styles.wrapper}>
