+++ src/hooks/useLogout.tsx
@@ -3,7 +3,7 @@ import { ImportantMessageSetterContext } from "../context/importantMessage";
 import { Alert } from "react-native";
 import { AuthStoreContext } from "../context/authStore";
 import { CampaignConfigsStoreContext } from "../context/campaignConfigsStore";
-// import { useNavigation } from "@react-navigation/native";
+import { useNavigation } from "@react-navigation/native";
 
 type AlertProps = {
   title: string;
@@ -20,7 +20,7 @@ export const useLogout = (): LogoutHook => {
   const { clearAuthCredentials } = useContext(AuthStoreContext);
   const { clearCampaignConfigs } = useContext(CampaignConfigsStoreContext);
   const [isLoggingOut, setIsLoggingOut] = useState(false);
-  // const navigation = useNavigation();
+  const navigation = useNavigation();
 
   const logout: LogoutHook["logout"] = useCallback(
     async alert => {
@@ -29,7 +29,7 @@ export const useLogout = (): LogoutHook => {
       clearCampaignConfigs();
       setMessageContent(null);
       setIsLoggingOut(false);
-      // navigation.navigate("LoginScreen");
+      navigation.navigate("LoginScreen");
       if (alert) {
         const { title, description } = alert;
         // using react-native alerts here for now because there are cases where the user is currently on a (non-alert) modal
@@ -37,7 +37,7 @@ export const useLogout = (): LogoutHook => {
         Alert.alert(title, description);
       }
     },
-    [clearAuthCredentials, clearCampaignConfigs, setMessageContent]
+    [clearAuthCredentials, clearCampaignConfigs, navigation, setMessageContent]
   );
 
   return {
