+++ src/navigation/index.tsx
@@ -1,5 +1,4 @@
-import React, { ReactElement, useRef } from "react";
-import { NavigationContainer } from "@react-navigation/native";
+import React, { ReactElement } from "react";
 import { AuthStoreContextProvider } from "../context/authStore";
 import { ConfigContextProvider } from "../context/config";
 import { FontLoader } from "../components/FontLoader";
@@ -12,12 +11,8 @@ import { Providers } from "../context/composeProviders";
 import { DrawerContextProvider } from "../context/drawer";
 import { CampaignConfigsStoreContextProvider } from "../context/campaignConfigsStore";
 import { SafeAreaProvider } from "react-native-safe-area-context";
-import * as Linking from "expo-linking";
 
 const App = (): ReactElement => {
-  const prefix = Linking.makeUrl("/");
-  const navigationRef = useRef(null);
-
   return (
     <ErrorBoundary>
       <FontLoader>
@@ -33,14 +28,7 @@ const App = (): ReactElement => {
             SafeAreaProvider
           ]}
         >
-          <NavigationContainer
-            linking={{
-              prefixes: [prefix]
-            }}
-            ref={navigationRef}
-          >
-            <Content />
-          </NavigationContainer>
+          <Content />
         </Providers>
       </FontLoader>
     </ErrorBoundary>
